fixed compatibility issues between windows and linux

+++ TestFortran/irf_test.f90
@@ -6,7 +6,7 @@
 
     type (bmi_fastmech) :: m
     integer :: s, i
-    real :: time, time0, time1
+    double precision :: time, time0, time1
     character (len=BMI_MAX_UNITS_NAME) :: time_units
 
     write (*,"(a)",advance="no") "Initializing..."
@@ -41,7 +41,7 @@
     write (*,"(a)") "Update a fraction of a time step"
     s = m%get_current_time(time0)
     write (*,"(a30, f8.2)") "Start time: ", time0
-    s = m%update_frac(0.5)
+    s = m%update_frac(0.5d0)
     s = m%get_current_time(time1)
     write (*,"(a30, f8.2)") "Stop time: ", time1
